requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
idna==3.4
Pillow==9.5.0
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
